Add spring-boot-starter-validation in Boot 2.3 if validation is used #306
Since Boot 2.3 `spring-boot-starter-validation` is no longer a transitive
dependency of the web starters.

While the code will probably compile fine due to the presence of
`validation-api` which is added as part of the Boot 2.0
migration runtime errors may occur because no Bean Validation
implementation is present on the classpath anymore.

See https://github.com/spring-projects/spring-boot/wiki/Spring-Boot-2.3-Release-Notes#validation-starter-no-longer-included-in-web-starters