[779] Remove `StepBuilderFactory` parameters from @Bean-annotated methods
This is removed in SpringBoot 5.2 and the 5.0 recipe is already removing
@Autowired fields.

I thought I would have to do something like this for
`JobBuilderFactory`, but its recipe is already doing somethign like
this.  It's a little unclear why there's more testing in this
`StepBuilderFactory` case than there is for `JobBuilderFactory`.

Regardless, this class is removed, and I'm hoping this will solve many
cases for us (and others!).

 #779