Remove `StepBuilderFactory` parameters from `@Bean`-annotated methods (#780)
* [779] Remove `StepBuilderFactory` parameters from @Bean-annotated methods

This is removed in SpringBoot 5.2 and the 5.0 recipe is already removing
@Autowired fields.

I thought I would have to do something like this for
`JobBuilderFactory`, but its recipe is already doing somethign like
this.  It's a little unclear why there's more testing in this
`StepBuilderFactory` case than there is for `JobBuilderFactory`.

Regardless, this class is removed, and I'm hoping this will solve many
cases for us (and others!).

 #779

* [779] Use Tim's ListUtils#mapFirst option to remove whitespace..

from front of method argument list.

 #779

* Retain the original first prefix of parameters

* Remove explicit `final`

---------

Co-authored-by: Samuel Cox <sacox@paypal.com>
Co-authored-by: Tim te Beek <tim@moderne.io>